# server/docs/swagger.yaml
definitions:
  api.AdminInitRootUserReq:
    properties:
      email:
        description: 邮箱
        type: string
      name:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  api.AdminInitRootUserRes:
    properties:
      token:
        type: string
    type: object
  api.AdminRoleCreateReq:
    properties:
      code:
        description: 角色编码
        type: string
      desc:
        description: 描述
        type: string
      name:
        description: 角色名称
        type: string
    required:
    - code
    - name
    type: object
  api.AdminRoleCreateRes:
    properties:
      id:
        description: 角色 ID
        type: integer
    type: object
  api.AdminRoleDeleteReq:
    properties:
      id:
        description: 角色 ID
        type: integer
    required:
    - id
    type: object
  api.AdminRoleListReq:
    properties:
      current:
        minimum: 1
        type: integer
      keyword:
        description: 查询关键字
        type: string
      order_key:
        type: string
      order_type:
        enum:
        - ase
        - desc
        type: string
      page_size:
        minimum: 0
        type: integer
    type: object
  api.AdminRoleListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AdminRole'
        type: array
      total:
        type: integer
    type: object
  api.AdminRoleUpdateCodesReq:
    properties:
      codes:
        description: 权限码
        items:
          type: string
        type: array
      id:
        description: 角色 ID
        type: integer
    required:
    - codes
    - id
    type: object
  api.AdminRoleUpdateReq:
    properties:
      code:
        description: 角色编码
        type: string
      desc:
        description: 描述
        type: string
      id:
        description: 角色 ID
        type: integer
      name:
        description: 角色名称
        type: string
    required:
    - code
    - id
    - name
    type: object
...
  api.AdminUserForgetPasswordReq:
    properties:
      captcha:
        description: 邮箱验证码
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
    required:
    - captcha
    - email
    - password
    type: object
...
  api.AdminUserLoginReq:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  api.AdminUserLoginRes:
    properties:
      token:
        type: string
    type: object
...
  consts.AdminUserStatus:
    enum:
    - 1
    - -1
    type: integer
    x-enum-comments:
      AdminUserStatusLocked: 锁定
      AdminUserStatusNormal: 正常
    x-enum-varnames:
    - AdminUserStatusNormal
    - AdminUserStatusLocked
...
  controller.loginBodyDto:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  controller.loginSuccessResponse:
    properties:
      token:
        type: string
    type: object
...
  model.AdminRole:
    properties:
      code:
        description: 角色编号
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        type: string
      id:
        description: ID
        type: integer
      name:
        type: string
      permission_codes:
        description: 权限码
        items:
          type: string
        type: array
      updated_at:
        description: 更新时间
        type: string
      users:
        items:
          $ref: '#/definitions/model.AdminUser'
        type: array
    required:
    - code
    - created_at
    - id
    - name
    - users
    type: object
  model.AdminUser:
    properties:
      avatar:
        type: string
      created_at:
        description: 创建时间
        type: string
      email:
        type: string
      id:
        description: ID
        type: integer
      is_root:
        description: 是否是超级管理员
        type: boolean
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/model.News'
        type: array
      roles:
        items:
          $ref: '#/definitions/model.AdminRole'
        type: array
      status:
        $ref: '#/definitions/consts.AdminUserStatus'
      updated_at:
        description: 更新时间
        type: string
      username:
        type: string
    required:
    - created_at
    - id
    - name
    - status
    - username
    type: object
  model.News:
    properties:
      author_id:
        description: 作者
        type: integer
      content:
        description: 内容
        type: string
      cover:
        description: 封面
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 简介
        type: string
      id:
        description: ID
        type: integer
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期
        type: string
      recommend:
        description: 推荐等级
        type: integer
      title:
        description: 新闻标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    required:
    - author_id
    - created_at
    - id
    - is_visible
    - recommend
    - title
    type: object
...
  resp.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: 男生自用 API 接口文档
  version: "1.0"
paths:
  /api/admin/auth/forget-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserForgetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 忘记密码
      tags:
      - 管理后台-用户认证
  /api/admin/auth/init-root-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminInitRootUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminInitRootUserRes'
              type: object
      summary: 初始化超级管理员用户
      tags:
      - 管理后台-用户认证
  /api/admin/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminUserLoginRes'
              type: object
      summary: 用户登录
      tags:
      - 管理后台-用户认证
...
  /api/admin/role/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminRoleCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminRoleCreateRes'
              type: object
      summary: 新增管理员角色
      tags:
      - 管理后台-管理员角色
  /api/admin/role/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminRoleDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 删除角色
      tags:
      - 管理后台-管理员角色
  /api/admin/role/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminRoleListReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminRoleListRes'
              type: object
      summary: 管理员角色列表
      tags:
      - 管理后台-管理员角色
...
  /api/admin/role/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminRoleUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改管理员角色
      tags:
      - 管理后台-管理员角色
  /api/admin/role/update/permission-codes:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminRoleUpdateCodesReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改管理员角色权限
      tags:
      - 管理后台-管理员角色
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.loginBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/controller.loginSuccessResponse'
              type: object
      summary: 用户登录
      tags:
      - 前台-用户
...
swagger: "2.0"